multi_tool_agent/requirements.txt
# Core dependencies
requests==2.31.0
python-dotenv==1.0.1
google-adk==0.0.1 